Moved tests to the appropriate package, a lot of indoor tests were inside the outdoor folder.
Fixed broken tests within ConcreteBuildingTest.

Related to issue #93, #31
