+++ app/src/androidTest/java/com/conupods/IndoorMaps/indoorPathTestClass.java
@@ -1,4 +1,4 @@
-package com.conupods;
+package com.conupods.IndoorMaps;
 
 
 import org.junit.Test;
